Fix a puppet-lint error
puppet-lint reported an error: "ERROR: single quoted string
containing a variable found". With this change it only reports a
warning: "double quoted string containing no variables".

There is no way to get rid of the warning with puppet-lint <1.0.0
(see https://github.com/rodjek/puppet-lint/issues/256 for details)
but it will be gone as soon as we upgrade to puppet-lint >=1.0.0.