Work around a problem with "make --question" in EPICS Base
A bug in EPICS base causes EPICS IOC applications not to handle
"make --question" correctly. The bug causes "make --question"
to always return "needs to be rebuild". See
https://bugs.launchpad.net/epics-base/+bug/1669891 for details.
This bug might cause IOCs to get rebuild and restarted on every
Puppet run (if configured to do so). This work around should fix
the problem by bypassing the broken code in the Makefiles and
thus fixes [DEPLOY-11].